+++ packages/core/src/impl-address.ts
@@ -1,12 +1,11 @@
-import { keccak256 } from 'ethereumjs-util';
 import {
-  call,
   EIP1967BeaconNotFound,
   EIP1967ImplementationNotFound,
   getBeaconAddress,
   getImplementationAddress,
   UpgradesError,
 } from '.';
+import { callOptionalSignature } from './call-optional-signature';
 
 import { EthereumProvider } from './provider';
 import { parseAddress } from './utils/address';
@@ -24,27 +23,17 @@ export async function getImplementationAddressFromBeacon(
   provider: EthereumProvider,
   beaconAddress: string,
 ): Promise<string> {
-  const implementationFunction = '0x' + keccak256(Buffer.from('implementation()')).toString('hex').slice(0, 8);
-  let result: string | undefined;
-  try {
-    const implAddress = await call(provider, beaconAddress, implementationFunction);
-    result = parseAddress(implAddress);
-  } catch (e: any) {
-    if (
-      !(
-        e.message.includes('function selector was not recognized') ||
-        e.message.includes('invalid opcode') ||
-        e.message.includes('revert') ||
-        e.message.includes('execution error')
-      )
-    ) {
-      throw e;
-    } // otherwise fall through with no result
+  const impl = await callOptionalSignature(provider, beaconAddress, 'implementation()');
+  let parsedImplAddress;
+  if (impl !== undefined) {
+    parsedImplAddress = parseAddress(impl);
   }
-  if (result === undefined) {
+
+  if (parsedImplAddress === undefined) {
     throw new InvalidBeacon(`Contract at ${beaconAddress} doesn't look like a beacon`);
+  } else {
+    return parsedImplAddress;
   }
-  return result;
 }
 
 /**