+++ packages/plugin-hardhat/contracts/Greeter50Proxiable.sol
@@ -1,0 +1,6 @@
+pragma solidity >= 0.4.22 <0.8.0;
+
+import "./Greeter.sol";
+import "./utils/Proxiable50.sol";
+
+contract Greeter50Proxiable is Greeter, Proxiable50 {}