+++ packages/plugin-hardhat/contracts/Import50.sol
@@ -1,0 +1,8 @@
+// SPDX-License-Identifier: MIT
+pragma solidity ^0.8.20;
+
+import "@openzeppelin/contracts-5.0/proxy/beacon/BeaconProxy.sol";
+import "@openzeppelin/contracts-5.0/proxy/beacon/UpgradeableBeacon.sol";
+import "@openzeppelin/contracts-5.0/proxy/ERC1967/ERC1967Proxy.sol";
+import "@openzeppelin/contracts-5.0/proxy/transparent/TransparentUpgradeableProxy.sol";
+import "@openzeppelin/contracts-5.0/proxy/transparent/ProxyAdmin.sol";